Adjusted MongoMovieService to check database size

+++ backend/src/main/kotlin/com/pk/zbtz/zbtzbackend/domain/Statistics.kt
@@ -2,5 +2,5 @@ package com.pk.zbtz.zbtzbackend.domain
 
 data class Statistics(
     val accessTime: Long = 0,
-    val databaseMemorySize: Long = 0,
+    val databaseMemorySize: Double = 0.0,
 )
